backend/requirements.txt
Django==4.2.7
django-crispy-forms==2.3
crispy-bootstrap4==2024.10
Pillow==10.1.0
anthropic==0.21.3
httpx==0.27.2
typing-extensions>=4.11.0
python-dotenv==1.0.0
openai==1.59.5
djangorestframework==3.15.2
django-cors-headers==4.6.0
djangorestframework-simplejwt==5.4.0
